+++ lua/gutenberg/constructs/code_block.lua
@@ -9,7 +9,7 @@
 
 local buffer = require('gutenberg.buffer')
 local context = require('gutenberg.context')
-local ts = require('gutenberg.ts')
+local ts = require('gutenberg.treesitter')
 
 ---@class gutenberg.code_block
 local M = {}