+++ lua/gutenberg/constructs/heading.lua
@@ -4,7 +4,7 @@
 
 local buffer = require('gutenberg.buffer')
 local context = require('gutenberg.context')
-local ts = require('gutenberg.ts')
+local ts = require('gutenberg.treesitter')
 
 ---@class gutenberg.heading
 local M = {}