+++ lua/gutenberg/constructs/table.lua
@@ -7,7 +7,7 @@
 
 local buffer = require('gutenberg.buffer')
 local context = require('gutenberg.context')
-local ts = require('gutenberg.ts')
+local ts = require('gutenberg.treesitter')
 
 ---@class gutenberg.table
 local M = {}